+++ packages/expect/src/__tests__/assertionCounts.test.ts
@@ -37,6 +37,7 @@ describe('.hasAssertions()', () => {
 
   it('throws if expected is not undefined', () => {
     jestExpect(() => {
+      // @ts-expect-error
       jestExpect.hasAssertions(2);
     }).toThrowErrorMatchingSnapshot();
   });