+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Fixes
 
-- `[jest-runtime]` Ensure absolute paths can be resolved within test modules ([11943](https://github.com/facebook/jest/pull/11943))
+- `[expect]` Tweak and improve types ([#11949](https://github.com/facebook/jest/pull/11949))
+- `[jest-runtime]` Ensure absolute paths can be resolved within test modules ([#11943](https://github.com/facebook/jest/pull/11943))
 
 ### Chore & Maintenance