+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 ### Chore & Maintenance
+- `[jest-config, jest-util]` Use `ci-info` instead of `is-ci` to detect CI environment ([11973](https://github.com/facebook/jest/pull/11973))
 
 ### Performance