+++ CHANGELOG.md
@@ -4,11 +4,11 @@
 
 ### Fixes
 
-- `[expect]` Make expect extension properties `configurable` ([11978](https://github.com/facebook/jest/pull/11978))
+- `[expect]` Make expect extension properties `configurable` ([#11978](https://github.com/facebook/jest/pull/11978))
 
 ### Chore & Maintenance
 
-- `[jest-config, jest-util]` Use `ci-info` instead of `is-ci` to detect CI environment ([11973](https://github.com/facebook/jest/pull/11973))
+- `[jest-config, jest-util]` Use `ci-info` instead of `is-ci` to detect CI environment ([#11973](https://github.com/facebook/jest/pull/11973))
 
 ### Performance
 
@@ -16,7 +16,7 @@
 
 ### Features
 
-- `[jest-config]` Add `testEnvironmentOptions.html` to apply to jsdom input ([11950](https://github.com/facebook/jest/pull/11950))
+- `[jest-config]` Add `testEnvironmentOptions.html` to apply to jsdom input ([#11950](https://github.com/facebook/jest/pull/11950))
 - `[jest-resolver]` Support default export (`.`) in `exports` field _if_ `main` is missing ([#11919](https://github.com/facebook/jest/pull/11919))
 
 ### Fixes